+++ NEWS.md
@@ -120,10 +120,10 @@ argument systematically anymore
 * The title is automatically filled with y.label argument
 * Add the remove.missing argument in report.quali to be able to compute the percentage on non missing values only if needed
 * Add transpose function
-
+* Add hierarchical descriptive statistics for reporting Adverse Event, Medical History and concomitant treatment
 
 ## Issues
-* Correction of a bug in report.quanti when x1 and x2 arguments weere NULL 
+* Correction of a bug in report.quanti when x1 and x2 arguments were NULL